backend/requirements.txt
flask
flask-cors
torch
torchvision
opencv-python
numpy
Pillow
gunicorn